[refactor] 분리한 클래스 다시 합치기
