--- src/main/java/com/jaeyeonling/calculate/utils/StringUtils.java
@@ -1,10 +1,0 @@
-package com.jaeyeonling.calculate.utils;
-
-public final class StringUtils {
-
-    private StringUtils() { }
-
-    public static boolean isNullOrBlank(final String source) {
-        return source == null || source.isBlank();
-    }
-}
